Set cursor color to make sure it is visible with current theme.
I was consistently losing my mouse pointer when switching between
monitors. This sets the mouse cursor to the color also used for the text
cursor. This is especially a problem when notification, which need mouse
input, come up on main screen (running full screen emacs) and I have no
idea where the mouse cursor is to click the notification away.